--- HelloWorld.java
@@ -1,6 +1,0 @@
-public class HelloWorld {
-
-    public String hello(){
-        return "Hello World!";
-    }
-}
