+++ LeetCode/Array/String/BasicsWithString.java
@@ -1,7 +1,4 @@
-package Strings;
-
-import java.util.ArrayList;
-import java.util.List;
+package LeetCode.Array.String;
 
 public class BasicsWithString {
